+++ algorithm/include/ChainedHashMapOLD.h
@@ -2,7 +2,7 @@
 
 #include "VComparator.h"
 #include "UniversalHash.h"
-#include "LinkedHashNode.h"
+#include "ChainedHashNode.h"
 
 
 // Hash map class template
@@ -42,7 +42,7 @@ public:
 
     bool search(const K& key, V& value)
     {
-        unsigned long hashValue = hashFunc[key];
+        unsigned int hashValue = hashFunc[key];
         HashNode<K, V>* entry = table[hashValue];
         unsigned int ccounter = 0;
         while (entry != NULL) {
@@ -63,7 +63,7 @@ public:
 
     void insert(const K& key, const V& value)
     {
-        unsigned long hashValue = hashFunc[key];
+        unsigned int hashValue = hashFunc[key];
         HashNode<K, V>* prev = NULL;
         HashNode<K, V>* entry = table[hashValue];
         // std::cout << "before while" << entry <<std::endl;
@@ -93,7 +93,7 @@ public:
 
     void remove(const K& key)
     {
-        unsigned long hashValue = hashFunc[key];
+        unsigned int hashValue = hashFunc[key];
         HashNode<K, V>* prev = NULL;
         HashNode<K, V>* entry = table[hashValue];
